moved buttons arround, some buttons display dynamically, fixed ROI position bug. Hardcoded version display temporarily. hardcoded python support for 3.6 only (hopefully temporarily). Changed installation instructions in READMe.md.
