+++ README.md
@@ -7,10 +7,10 @@ To be completed
 Pre-requisites
 ==============
 
-Install Anaconda and create a python 3.7 conda environment:
+Install Anaconda3 and create a python 3.6 conda environment:
 
-    $ conda create -n py37 python=3.7
-    $ source activate py37
+    $ conda create -n py36 python=3.6
+    $ source activate py36
     
 add conda channels:
 
@@ -24,29 +24,6 @@ Install with:
 
     $ conda install -c fabriciosm xrftomo 
 
-OR install from source along with all dependencies:
-
-    $ git clone https://github.com/FabricioSMarin/xrftomo.git
-    $ cd xrftomo
-    $ conda install -c conda-forge dxchange
-    $ conda install -c conda-forge tomopy
-    $ conda install -c conda-forge pyqt
-    $ conda install -c conda-forge h5py
-    $ conda install -c conda-forge scikit-image
-    $ conda install -c conda-forge pandas
-    $ conda install -c conda-forge seaborn
-    $ conda install -c conda-forge pyqtgraph
-    $ conda install -c conda-forge scipy
-    $ conda install git
-    $ conda install numpy
-
-
-    $ python setup.py install
-    $ cd bin 
-    $ python xrftomo init
-
-in the prepared virtualenv or as root for system-wide installation.
-
 Usage
 =====
 
@@ -56,8 +33,8 @@ You can run xrftomo using a command line, for a full list of options use:
 
 or a GUI with:
 
-    $ bin/xrftomo gui
-    $ or double-click on xrftomo.sh
+    $ xrftomo gui
+    $ xrftomo gui --experimental
 
 ```
 If your python installation is in a location different from #!/usr/bin/env python please edit the first line of the bin/xrftomo file to match yours.
