+++ xrftomo/widgets/image_process_controls.py
@@ -65,8 +65,8 @@ class ImageProcessControlsWidget(QtWidgets.QWidget):
         self.combo1.setFixedWidth(button1size)
         self.combo2 = QtWidgets.QComboBox()
         self.combo2.setFixedWidth(button1size)
-        self.combo3 = QtWidgets.QComboBox(self)
-        self.combo3.setFixedWidth(button2size)
+        # self.combo3 = QtWidgets.QComboBox(self)
+        # self.combo3.setFixedWidth(button2size)
 
         # self.xSizeLbl = QtWidgets.QLabel("x Size")
         # self.xSizeLbl.setFixedWidth(button4size)
@@ -95,38 +95,35 @@ class ImageProcessControlsWidget(QtWidgets.QWidget):
         self.reshapeBtn.setFixedWidth(button33size)
         self.cropBtn = QtWidgets.QPushButton("Crop")
         self.cropBtn.setFixedWidth(button33size)
-        self.captureBackground = QtWidgets.QPushButton("Copy Bg")
+        self.captureBackground = QtWidgets.QPushButton("Copy")
         self.captureBackground.setFixedWidth(button33size)
-        self.setBackground = QtWidgets.QPushButton("Set Bg")
+        self.setBackground = QtWidgets.QPushButton("Paste")
         self.setBackground.setFixedWidth(button33size)
         self.deleteProjection = QtWidgets.QPushButton("remove img")
-        self.deleteProjection.setFixedWidth(button33size)
+        self.deleteProjection.setFixedWidth(button2size)
         # self.hist_equalize = QtWidgets.QPushButton("Equalize")
         # self.hist_equalize.setFixedWidth(button33size)
-        self.rm_hotspot = QtWidgets.QPushButton("Remove hs")
-        self.rm_hotspot.setFixedWidth(button33size)
+        self.rm_hotspot = QtWidgets.QPushButton("Remove hotspot")
+        self.rm_hotspot.setFixedWidth(button2size)
         self.Equalize = QtWidgets.QPushButton("Equalize")
         self.Equalize.setFixedWidth(button33size)
-        self.rot_axis = QtWidgets.QPushButton("Set rot. axis")
-        self.rot_axis.setFixedWidth(button33size)
 
 
-        for i in range(5):
-            self.combo3.addItem(str(i + 1))
-        self.btn1 = QtWidgets.QPushButton("fit to a line")
-        self.btn1.setFixedWidth(button2size)
-        self.btn2 = QtWidgets.QPushButton("fit to sine curve")
-        self.btn2.setFixedWidth(button2size)
-        self.btn3 = QtWidgets.QPushButton("set y")
-        self.btn3.setFixedWidth(button2size)
-        self.btn4 = QtWidgets.QPushButton("Clear data")
-        self.btn4.setFixedWidth(button2size)
-        self.lbl3 = QtWidgets.QLabel("hotspot group#")
-        self.lbl3.setFixedWidth(button2size)
+        # for i in range(5):
+        #     self.combo3.addItem(str(i + 1))
+        # self.btn1 = QtWidgets.QPushButton("fit to a line")
+        # self.btn1.setFixedWidth(button2size)
+        # self.btn2 = QtWidgets.QPushButton("fit to sine curve")
+        # self.btn2.setFixedWidth(button2size)
+        # self.btn3 = QtWidgets.QPushButton("set y")
+        # self.btn3.setFixedWidth(button2size)
+        # self.btn4 = QtWidgets.QPushButton("Clear data")
+        # self.btn4.setFixedWidth(button2size)
+        # self.lbl3 = QtWidgets.QLabel("hotspot group#")
+        # self.lbl3.setFixedWidth(button2size)
 
         self.Equalize.setVisible(False)
         self.reshapeBtn.setVisible(False)
-        self.btn2.setVisible(False)
 
         # hb1 = QtWidgets.QHBoxLayout()
         # hb1.addWidget(self.xSizeLbl)
@@ -137,21 +134,20 @@ class ImageProcessControlsWidget(QtWidgets.QWidget):
         # hb2.addWidget(self.y_sld)
         # hb2.addWidget(self.ySizeTxt)
 
-        hb3 = QtWidgets.QHBoxLayout()
-        hb3.addWidget(self.lbl3)
-        hb3.addWidget(self.combo3)
-        hb4 = QtWidgets.QHBoxLayout()
-        hb4.addWidget(self.btn1)
-        hb4.addWidget(self.btn3)        
-        hb5 = QtWidgets.QHBoxLayout()
-        hb5.addWidget(self.btn2)
-        hb5.addWidget(self.btn4)
+        # hb3 = QtWidgets.QHBoxLayout()
+        # hb3.addWidget(self.lbl3)
+        # hb3.addWidget(self.combo3)
+        # hb4 = QtWidgets.QHBoxLayout()
+        # hb4.addWidget(self.btn1)
+        # hb4.addWidget(self.btn3)        
+        # hb5 = QtWidgets.QHBoxLayout()
+        # hb5.addWidget(self.btn2)
+        # hb5.addWidget(self.btn4)
 
         hb10 = QtWidgets.QHBoxLayout()
         # hb10.addWidget(self.normalizeBtn)
         # hb10.addWidget(self.hist_equalize)
         hb10.addWidget(self.Equalize)
-        hb10.addWidget(self.rot_axis)
 
 
         hb12 = QtWidgets.QHBoxLayout()
@@ -168,10 +164,10 @@ class ImageProcessControlsWidget(QtWidgets.QWidget):
         # vb1.addLayout(hb1)
         # vb1.addLayout(hb2)
 
-        vb2 = QtWidgets.QVBoxLayout()
-        vb2.addLayout(hb3)
-        vb2.addLayout(hb4)
-        vb2.addLayout(hb5)
+        # vb2 = QtWidgets.QVBoxLayout()
+        # vb2.addLayout(hb3)
+        # vb2.addLayout(hb4)
+        # vb2.addLayout(hb5)
 
         vb4 = QtWidgets.QVBoxLayout()
         vb4.addLayout(hb10)
@@ -185,7 +181,7 @@ class ImageProcessControlsWidget(QtWidgets.QWidget):
         vb5.addWidget(self.combo2)
         vb5.addLayout(vb1)
         vb5.addLayout(vb4)
-        vb5.addLayout(vb2)
+        # vb5.addLayout(vb2)
 
         self.setLayout(vb5)
